% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_optimization.R
\name{expected_improvement}
\alias{expected_improvement}
\title{Title}
\usage{
expected_improvement(X, X_sample, Y_sample, gpr, xi = 0.01, max = TRUE)
}
\arguments{
\item{X}{}

\item{X_sample}{}

\item{Y_sample}{}

\item{gpr}{}

\item{xi}{}

\item{max}{}
}
\value{

}
\description{
Title
}
