% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Title}
\usage{
gaussian_kernel(x, y, l = 1, sigma_f = 1)
}
\arguments{
\item{x}{}

\item{y}{}

\item{l}{}

\item{sigma_f}{}
}
\value{

}
\description{
Title
}
