% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{gpr.init}
\alias{gpr.init}
\title{Title}
\usage{
gpr.init(l = 1, sigma_f = 1, sigma_y = 0, kern = gaussian_kernel)
}
\arguments{
\item{l}{}

\item{sigma_f}{}

\item{sigma_y}{}

\item{kern}{}
}
\value{

}
\description{
Title
}
