% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{len}
\alias{len}
\title{Generalized length function}
\usage{
len(x)
}
\arguments{
\item{x}{a vector or a two-dimensional array.}
}
\value{
an integer of length 1 or NULL.
}
\description{
Returns the length of \code{x} if it's a vector or the number of rows if \code{x} is an array.
}
\examples{
len(c(1,2,3))
len(diag(4))
}
