% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{nll_fn}
\alias{nll_fn}
\title{Title}
\usage{
nll_fn(X_train, Y_train, noise, kernel)
}
\arguments{
\item{X_train}{}

\item{Y_train}{}

\item{noise}{}

\item{kernel}{}
}
\value{

}
\description{
Title
}
