% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{normalize}
\alias{normalize}
\title{Row-Wise Matrix Normalization}
\usage{
normalize(X)
}
\arguments{
\item{X}{a two-dimensional \code{m * n} numeric array.}
}
\value{
an \code{m * n} normalized numeric array.
}
\description{
Normalize a two-dimensional numeric array such that its row sums add up to 1.
}
\examples{
a <- array(c(1,4,2,5,3,6),c(2,3))
normalize(a)
}
