% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_optimization.R
\name{propose_location}
\alias{propose_location}
\title{Title}
\usage{
propose_location(
  acquisition,
  X_sample,
  Y_sample,
  gpr,
  lower,
  upper,
  n_restarts = 25,
  xi = 0.01,
  max = TRUE
)
}
\arguments{
\item{acquisition}{}

\item{X_sample}{}

\item{Y_sample}{}

\item{gpr}{}

\item{lower}{}

\item{upper}{}

\item{n_restarts}{}

\item{xi}{}

\item{max}{}
}
\value{

}
\description{
Title
}
