% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{rappend}
\alias{rappend}
\title{Generalized append function}
\usage{
rappend(X, Y)
}
\arguments{
\item{X}{the vector or two-dimensional matrix to be appended to}

\item{Y}{the row vector or numeric value to be added}
}
\value{
A vector containing the values in \code{X} with \code{Y} appended
at the end of the vector or a matrix \code{X} with the row vector
\code{Y} added to the bottom.
}
\description{
Appends a row vector to a matrix or an element to a vector.
In the case of two vectors, stack them on top of each other
to form a two-dimensional matrix.
}
\examples{
rappend(c(1,2,3), 4)
rappend(diag(3), c(1,2,3))
rappend(c(1,2,3), c(4,5,6))
}
