% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_optimization.R
\name{bayesian_optimization}
\alias{bayesian_optimization}
\title{Title}
\usage{
bayesian_optimization(
  FUN,
  lower,
  upper,
  init_grid_dt = NULL,
  init_points = 0,
  n_iter = 10,
  xi = 0.01,
  noise = 0,
  max = TRUE,
  acq = expected_improvement,
  naive = FALSE
)
}
\arguments{
\item{FUN}{}

\item{lower}{}

\item{upper}{}

\item{init_grid_dt}{}

\item{init_points}{}

\item{n_iter}{}

\item{xi}{}

\item{noise}{}

\item{max}{}

\item{acq}{}

\item{naive}{}
}
\value{

}
\description{
Title
}
