% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{gpr.fit}
\alias{gpr.fit}
\title{Title}
\usage{
gpr.fit(
  X_train,
  Y_train,
  gpr,
  lower = c(1e-05, 1e-05),
  upper = c(2, 2),
  n_restarts = 25,
  naive = FALSE
)
}
\arguments{
\item{X_train}{}

\item{Y_train}{}

\item{gpr}{}

\item{lower}{}

\item{upper}{}

\item{n_restarts}{}

\item{naive}{}
}
\value{

}
\description{
Title
}
