% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_optimization.R
\name{plot_acquisition}
\alias{plot_acquisition}
\title{Title}
\usage{
plot_acquisition(X, EI, X_next = NULL)
}
\arguments{
\item{X}{}

\item{EI}{}

\item{X_next}{}
}
\value{

}
\description{
Title
}
