% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{plot_gp}
\alias{plot_gp}
\title{Title}
\usage{
plot_gp(mu, cov, X, X_train = NULL, Y_train = NULL, samples = NULL)
}
\arguments{
\item{mu}{}

\item{cov}{}

\item{X}{}

\item{X_train}{}

\item{Y_train}{}

\item{samples}{}
}
\value{

}
\description{
Title
}
